# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeasureBody": {
        "properties": {
          "on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On"
          },
          "values": {
            "additionalProperties": {
              "type": "number"
            },
            "maxProperties": 6,
            "propertyNames": {
              "enum": [
                "bodyweight",
                "waist",
                "chest",
                "upper_arm",
                "thigh",
                "resting_hr"
              ]
            },
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "values"
        ],
        "title": "MeasureBody",
        "type": "object"
      },
      "MeasurementView": {
        "properties": {
          "kind": {
            "enum": [
              "bodyweight",
              "waist",
              "chest",
              "upper_arm",
              "thigh",
              "resting_hr"
            ],
            "title": "Kind",
            "type": "string"
          },
          "on": {
            "format": "date",
            "title": "On",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "on",
          "kind",
          "value"
        ],
        "title": "MeasurementView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Training Coach",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/body": {
      "get": {
        "description": "Every measurement, newest day first.",
        "operationId": "history_api_body_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MeasurementView"
                  },
                  "title": "Response History Api Body Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "History",
        "tags": [
          "body"
        ]
      },
      "put": {
        "description": "Save a day's measurements (any of them), replacing the same kind that day.",
        "operationId": "save_api_body_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeasureBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save",
        "tags": [
          "body"
        ]
      }
    },
...
    "/api/body/{on}/{kind}": {
      "delete": {
        "operationId": "remove_api_body__on___kind__delete",
        "parameters": [
          {
            "in": "path",
            "name": "on",
            "required": true,
            "schema": {
              "format": "date",
              "title": "On",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "bodyweight",
                "waist",
                "chest",
                "upper_arm",
                "thigh",
                "resting_hr"
              ],
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove",
        "tags": [
          "body"
        ]
      }
    },
...
  }
}